fix: Adding protected routes
